Increase decrease working
Now when clicking on the skill increase button on an SR5 system Actor it will atempt to increase or decrease the value, that will add a history log to the actor as a flag and it will log it in a journal. This should also include purchases from items and sold items once supported by the modul.