GlobalStorageSelectedToken
Now we can add an item to the basket for a selected actor as GM, the player will have to assign an actor to himself to be able to add items to the basket. right now the adding will work as it still uses the actor flag. Remove now also works, next up is update