AvailabilityContext in inGameMarketplace
This now contains a First inGameMarketplace Dialog box for the Availability test.

It adds standalone Availability roll that we can link in later on the SR5Actor sheet. It adds a helper dialogList, this helper provides a list of dialog options for the AvailabilityTest.
AppDialogBuilder will build the Dialog based on the current progress in the Availability testing, this currently covers the initial build and the build for the first test. Added Localisation for skills as taking the system locals was an issue, looking for a way to remove them again...